+++ web/tests/typing_status.test.cjs
@@ -64,12 +64,20 @@ run_test("basics", ({override, override_rewire}) => {
     set_global("clearTimeout", clear_timeout);
 
     function notify_server_start(recipient) {
-        assert.deepStrictEqual(recipient, {message_type: "direct", ids: [1, 2]});
+        assert.deepStrictEqual(recipient, {
+            message_type: "direct",
+            notification_event_type: "typing",
+            ids: [1, 2],
+        });
         events.started = true;
     }
 
     function notify_server_stop(recipient) {
-        assert.deepStrictEqual(recipient, {message_type: "direct", ids: [1, 2]});
+        assert.deepStrictEqual(recipient, {
+            message_type: "direct",
+            notification_event_type: "typing",
+            ids: [1, 2],
+        });
         events.stopped = true;
     }
 
@@ -97,11 +105,11 @@ run_test("basics", ({override, override_rewire}) => {
     };
 
     // Start talking to users having ids - 1, 2.
-    call_handler({message_type: "direct", ids: [1, 2]});
+    call_handler({message_type: "direct", notification_event_type: "typing", ids: [1, 2]});
     assert.deepEqual(typing_status.state, {
         next_send_start_time: make_time(5 + 10),
         idle_timer: "idle_timer_stub",
-        current_recipient: {message_type: "direct", ids: [1, 2]},
+        current_recipient: {message_type: "direct", notification_event_type: "typing", ids: [1, 2]},
     });
     assert.deepEqual(events, {
         idle_callback: events.idle_callback,
@@ -113,11 +121,11 @@ run_test("basics", ({override, override_rewire}) => {
 
     // type again 3 seconds later
     worker.get_current_time = returns_time(8);
-    call_handler({message_type: "direct", ids: [1, 2]});
+    call_handler({message_type: "direct", notification_event_type: "typing", ids: [1, 2]});
     assert.deepEqual(typing_status.state, {
         next_send_start_time: make_time(5 + 10),
         idle_timer: "idle_timer_stub",
-        current_recipient: {message_type: "direct", ids: [1, 2]},
+        current_recipient: {message_type: "direct", notification_event_type: "typing", ids: [1, 2]},
     });
     assert.deepEqual(events, {
         idle_callback: events.idle_callback,
@@ -130,11 +138,11 @@ run_test("basics", ({override, override_rewire}) => {
     // type after 15 secs, so that we can notify the server
     // again
     worker.get_current_time = returns_time(18);
-    call_handler({message_type: "direct", ids: [1, 2]});
+    call_handler({message_type: "direct", notification_event_type: "typing", ids: [1, 2]});
     assert.deepEqual(typing_status.state, {
         next_send_start_time: make_time(18 + 10),
         idle_timer: "idle_timer_stub",
-        current_recipient: {message_type: "direct", ids: [1, 2]},
+        current_recipient: {message_type: "direct", notification_event_type: "typing", ids: [1, 2]},
     });
     assert.deepEqual(events, {
         idle_callback: events.idle_callback,
@@ -167,11 +175,11 @@ run_test("basics", ({override, override_rewire}) => {
 
     // Start talking to users again.
     worker.get_current_time = returns_time(50);
-    call_handler({message_type: "direct", ids: [1, 2]});
+    call_handler({message_type: "direct", notification_event_type: "typing", ids: [1, 2]});
     assert.deepEqual(typing_status.state, {
         next_send_start_time: make_time(50 + 10),
         idle_timer: "idle_timer_stub",
-        current_recipient: {message_type: "direct", ids: [1, 2]},
+        current_recipient: {message_type: "direct", notification_event_type: "typing", ids: [1, 2]},
     });
     assert.deepEqual(events, {
         idle_callback: events.idle_callback,
@@ -193,11 +201,11 @@ run_test("basics", ({override, override_rewire}) => {
 
     // Start talking to users again.
     worker.get_current_time = returns_time(80);
-    call_handler({message_type: "direct", ids: [1, 2]});
+    call_handler({message_type: "direct", notification_event_type: "typing", ids: [1, 2]});
     assert.deepEqual(typing_status.state, {
         next_send_start_time: make_time(80 + 10),
         idle_timer: "idle_timer_stub",
-        current_recipient: {message_type: "direct", ids: [1, 2]},
+        current_recipient: {message_type: "direct", notification_event_type: "typing", ids: [1, 2]},
     });
     assert.deepEqual(events, {
         idle_callback: events.idle_callback,
@@ -229,11 +237,11 @@ run_test("basics", ({override, override_rewire}) => {
 
     // Start talking to users again.
     worker.get_current_time = returns_time(170);
-    call_handler({message_type: "direct", ids: [1, 2]});
+    call_handler({message_type: "direct", notification_event_type: "typing", ids: [1, 2]});
     assert.deepEqual(typing_status.state, {
         next_send_start_time: make_time(170 + 10),
         idle_timer: "idle_timer_stub",
-        current_recipient: {message_type: "direct", ids: [1, 2]},
+        current_recipient: {message_type: "direct", notification_event_type: "typing", ids: [1, 2]},
     });
     assert.deepEqual(events, {
         idle_callback: events.idle_callback,
@@ -247,15 +255,19 @@ run_test("basics", ({override, override_rewire}) => {
     worker.get_current_time = returns_time(171);
 
     worker.notify_server_start = (recipient) => {
-        assert.deepStrictEqual(recipient, {message_type: "direct", ids: [3, 4]});
+        assert.deepStrictEqual(recipient, {
+            message_type: "direct",
+            notification_event_type: "typing",
+            ids: [3, 4],
+        });
         events.started = true;
     };
 
-    call_handler({message_type: "direct", ids: [3, 4]});
+    call_handler({message_type: "direct", notification_event_type: "typing", ids: [3, 4]});
     assert.deepEqual(typing_status.state, {
         next_send_start_time: make_time(171 + 10),
         idle_timer: "idle_timer_stub",
-        current_recipient: {message_type: "direct", ids: [3, 4]},
+        current_recipient: {message_type: "direct", notification_event_type: "typing", ids: [3, 4]},
     });
     assert.deepEqual(events, {
         idle_callback: events.idle_callback,
@@ -367,12 +379,22 @@ run_test("stream_messages", ({override, override_rewire}) => {
     set_global("clearTimeout", clear_timeout);
 
     function notify_server_start(recipient) {
-        assert.deepStrictEqual(recipient, {message_type: "stream", stream_id: 3, topic: "test"});
+        assert.deepStrictEqual(recipient, {
+            message_type: "stream",
+            notification_event_type: "typing",
+            stream_id: 3,
+            topic: "test",
+        });
         events.started = true;
     }
 
     function notify_server_stop(recipient) {
-        assert.deepStrictEqual(recipient, {message_type: "stream", stream_id: 3, topic: "test"});
+        assert.deepStrictEqual(recipient, {
+            message_type: "stream",
+            notification_event_type: "typing",
+            stream_id: 3,
+            topic: "test",
+        });
         events.stopped = true;
     }
 
@@ -400,11 +422,21 @@ run_test("stream_messages", ({override, override_rewire}) => {
     };
 
     // Start typing stream message
-    call_handler({message_type: "stream", stream_id: 3, topic: "test"});
+    call_handler({
+        message_type: "stream",
+        notification_event_type: "typing",
+        stream_id: 3,
+        topic: "test",
+    });
     assert.deepEqual(typing_status.state, {
         next_send_start_time: make_time(5 + 10),
         idle_timer: "idle_timer_stub",
-        current_recipient: {message_type: "stream", stream_id: 3, topic: "test"},
+        current_recipient: {
+            message_type: "stream",
+            notification_event_type: "typing",
+            stream_id: 3,
+            topic: "test",
+        },
     });
     assert.deepEqual(events, {
         idle_callback: events.idle_callback,
@@ -416,11 +448,21 @@ run_test("stream_messages", ({override, override_rewire}) => {
 
     // type again 3 seconds later. Covers 'same_stream_and_topic' codepath.
     worker.get_current_time = returns_time(8);
-    call_handler({message_type: "stream", stream_id: 3, topic: "test"});
+    call_handler({
+        message_type: "stream",
+        notification_event_type: "typing",
+        stream_id: 3,
+        topic: "test",
+    });
     assert.deepEqual(typing_status.state, {
         next_send_start_time: make_time(5 + 10),
         idle_timer: "idle_timer_stub",
-        current_recipient: {message_type: "stream", stream_id: 3, topic: "test"},
+        current_recipient: {
+            message_type: "stream",
+            notification_event_type: "typing",
+            stream_id: 3,
+            topic: "test",
+        },
     });
     assert.deepEqual(events, {
         idle_callback: events.idle_callback,
@@ -433,6 +475,188 @@ run_test("stream_messages", ({override, override_rewire}) => {
     // Explicitly stop.
     call_handler(null);
     assert.deepEqual(typing_status.state, null);
+    assert.deepEqual(events, {
+        idle_callback: undefined,
+        started: false,
+        stopped: true,
+        timer_cleared: true,
+    });
+});
+
+run_test("edit_messages", ({override_rewire}) => {
+    override_rewire(typing_status, "state", null);
+
+    let worker = {};
+    const events = {};
+    const message_id = 7;
+
+    function set_timeout(f, delay) {
+        assert.equal(delay, 5000);
+        events.idle_callback = f;
+        return "idle_timer_stub";
+    }
+
+    function clear_timeout() {
+        events.timer_cleared = true;
+    }
+
+    set_global("setTimeout", set_timeout);
+    set_global("clearTimeout", clear_timeout);
+
+    function notify_server_editing_start(recipient) {
+        assert.deepStrictEqual(recipient, {
+            notification_event_type: "typing_message_edit",
+            message_id,
+        });
+        events.started = true;
+    }
+
+    function notify_server_editing_stop(recipient) {
+        assert.deepStrictEqual(recipient, {
+            notification_event_type: "typing_message_edit",
+            message_id,
+        });
+        events.stopped = true;
+    }
+
+    function clear_events() {
+        events.idle_callback = undefined;
+        events.started = false;
+        events.stopped = false;
+        events.timer_cleared = false;
+    }
+
+    function call_handler_start(new_recipient) {
+        clear_events();
+        typing_status.update_editing_status(
+            worker,
+            new_recipient,
+            "start",
+            TYPING_STARTED_WAIT_PERIOD,
+            TYPING_STOPPED_WAIT_PERIOD,
+        );
+    }
+
+    function call_handler_stop(new_recipient) {
+        clear_events();
+        typing_status.update_editing_status(
+            worker,
+            new_recipient,
+            "stop",
+            TYPING_STARTED_WAIT_PERIOD,
+            TYPING_STOPPED_WAIT_PERIOD,
+        );
+    }
+
+    worker = {
+        get_current_time: returns_time(5),
+        notify_server_editing_start,
+        notify_server_editing_stop,
+    };
+
+    // Start typing stream message
+    call_handler_start({
+        notification_event_type: "typing_message_edit",
+        message_id,
+    });
+    assert.deepEqual(typing_status.editing_state.get(message_id), {
+        next_send_start_time: make_time(5 + 10),
+        idle_timer: "idle_timer_stub",
+        current_recipient: {
+            notification_event_type: "typing_message_edit",
+            message_id,
+        },
+    });
+    assert.deepEqual(events, {
+        idle_callback: events.idle_callback,
+        started: true,
+        stopped: false,
+        timer_cleared: false,
+    });
+    assert.ok(events.idle_callback);
+
+    worker.get_current_time = returns_time(8);
+    call_handler_start({
+        notification_event_type: "typing_message_edit",
+        message_id,
+    });
+    assert.deepEqual(typing_status.editing_state.get(message_id), {
+        next_send_start_time: make_time(5 + 10),
+        idle_timer: "idle_timer_stub",
+        current_recipient: {
+            notification_event_type: "typing_message_edit",
+            message_id,
+        },
+    });
+    assert.deepEqual(events, {
+        idle_callback: events.idle_callback,
+        started: false,
+        stopped: false,
+        timer_cleared: true,
+    });
+    assert.ok(events.idle_callback);
+
+    worker.get_current_time = returns_time(18);
+    call_handler_start({
+        notification_event_type: "typing_message_edit",
+        message_id,
+    });
+    assert.deepEqual(typing_status.editing_state.get(message_id), {
+        next_send_start_time: make_time(18 + 10),
+        idle_timer: "idle_timer_stub",
+        current_recipient: {
+            notification_event_type: "typing_message_edit",
+            message_id,
+        },
+    });
+    assert.deepEqual(events, {
+        idle_callback: events.idle_callback,
+        started: true,
+        stopped: false,
+        timer_cleared: true,
+    });
+    assert.ok(events.idle_callback);
+
+    // Now call recipients idle callback that we captured earlier.
+    const callback = events.idle_callback;
+    clear_events();
+    callback();
+    assert.deepEqual(typing_status.editing_state.get(message_id), undefined);
+    assert.deepEqual(events, {
+        idle_callback: undefined,
+        started: false,
+        stopped: true,
+        timer_cleared: true,
+    });
+
+    // Start editing message again.
+    worker.get_current_time = returns_time(50);
+    call_handler_start({
+        notification_event_type: "typing_message_edit",
+        message_id,
+    });
+    assert.deepEqual(typing_status.editing_state.get(message_id), {
+        next_send_start_time: make_time(50 + 10),
+        idle_timer: "idle_timer_stub",
+        current_recipient: {
+            notification_event_type: "typing_message_edit",
+            message_id,
+        },
+    });
+    assert.deepEqual(events, {
+        idle_callback: events.idle_callback,
+        started: true,
+        stopped: false,
+        timer_cleared: false,
+    });
+    assert.ok(events.idle_callback);
+
+    // Explicitly stop.
+    call_handler_stop({
+        notification_event_type: "typing_message_edit",
+        message_id,
+    });
+    assert.deepEqual(typing_status.editing_state.get(message_id), undefined);
     assert.deepEqual(events, {
         idle_callback: undefined,
         started: false,